Add files via upload
These are the files for the Aternos AFK bot. What the bot does is it joins your aternos server and makes it 24/7! If you need a tutorial here is the link- https://www.youtube.com/watch?v=Ux50ozXguUM&t=464s

Please watch the above video for help! Thank You!